+++ DemoApp/Sources/Views/CallView/CallingView/SimpleCallingView.swift
@@ -188,6 +188,15 @@ struct SimpleCallingView: View {
         try await call.updateAudioSessionPolicy(AppEnvironment.audioSessionPolicy.value)
     }
 
+    private func setProximityPolicies(for callId: String) throws {
+        let policies = AppEnvironment.proximityPolicies.map(\.value)
+        guard !policies.isEmpty else {
+            return
+        }
+        let call = streamVideo.call(callType: callType, callId: callId)
+        try policies.forEach { try call.addProximityPolicy($0) }
+    }
+
     private func parseURLIfRequired(_ text: String) {
         let adapter = DeeplinkAdapter()
         guard
@@ -224,6 +233,7 @@ struct SimpleCallingView: View {
         case .lobby:
             await setPreferredVideoCodec(for: text)
             try? await setAudioSessionPolicyOverride(for: text)
+            try? setProximityPolicies(for: text)
             viewModel.enterLobby(
                 callType: callType,
                 callId: text,
@@ -232,10 +242,12 @@ struct SimpleCallingView: View {
         case .join:
             await setPreferredVideoCodec(for: text)
             try? await setAudioSessionPolicyOverride(for: text)
+            try? setProximityPolicies(for: text)
             viewModel.joinCall(callType: callType, callId: text)
         case let .start(callId):
             await setPreferredVideoCodec(for: callId)
             try? await setAudioSessionPolicyOverride(for: callId)
+            try? setProximityPolicies(for: text)
             viewModel.startCall(
                 callType: callType,
                 callId: callId,